Add function to check if reg key exists
Added functionality which allows for users to check if a reg key
exists without first having to open and enumerate keys. This wouldn't
work anyway because opening keys that don't exist result in failures.

This new functionality just returns a boolean to indicate if the key
is present on the remote host.