Remove profiling block
MSP-8948

Believe that part of the profiling block, potentially the open, is
hanging the build on travis-ci, so remove the profiling block.
Additionally, the block shouldn't be in the final merged code anyway, it
just needs to be there to verify the timing changes, which can be
written to reference the previous commit.