Allow HttpClient to remember my cookie
When the target host sets a cookie, the Httpclient should remember
it. That way in the module I won't have to manually extract it,
and then put it in the next request.

If some for reason the module decides to send to different hosts,
the HttpClient mixin should know which one to use.