Speed up database usage in specs
MSP-9121

Instead of establishing a connection as needed for each access to the
database, maintain a background connection for the entire suite as is
done for rails specs.  This connection can be suspended by tagging the
example group with `:without_established_connection`, which will
remove the suite connection and re-establish it after all the examples
in the group have run.

This change brought the spec time down from 30 minutes to 12 minutes.