Remove Auto target from Firefox 17 rce.
It was always a lie and never worked for anything besides windows anyways :-/
This is because the listener must be decided and spun up at initialization time,
so we can't be waiting for the request to make that decision.

As a fix for this problem, we might consider spinning up one handler per target
at launch time, and dishing out the payload ports correctly. But that would take
some rewiring.

For now I will just disable auto targeting from this module so I stop forgetting
that it doesn't actually work.

Funny enough, symantech flags Windows x86 payloads *embedded* in an OSX macho
exe template, but misses actual osx payloads. Lol.