cmd_search
MSP-8986

Make cmd_search compatible with the new module cache.  Use TablePrint to
allow searched for columns to automatically appear in the table.  This
means when you search for an author or target platform you'll actually
be able to see them in the table, which wasn't the case previously.  To
give more control to the user added -d/--display to add columns to the
table that weren't even searched for and -D/--hide to hide columns that
are there by default or would be added because they are part of the
search.  Advanced tab completion for attribute and association
operators:  values will be plucked from database so operations as a
whole can be tab completed.