Fix tab completion for '.'
MSP-8986

Rex::Ui::Text::DispatcherShell#tab_complete_stub used Regexp instead of
the simpler String.start_with? to match tab completion items to the
partial word, which meant that '.' was treated as the single character
matcher instead of a literal '.'.  As there was an exception handler to
Regexp.escape the partial word if Regexp.compile failed, it is obvious
that actual Regexp support was not the purpose.