Fix #paired_real_path_sha1_hex_digests
MSP-9197

Due to stubbing without regard to the actual payload_type and the fact
that generating the metasploit_module involves generating a module_class
and module_ancestors for it, the number of module ancestors for a given
payload_type was variable.  Change the lets to remove the stub and
properly take into account the module ancestors generated by the module
class factory.