Quote table and column names for tab completion pluck
MSP-8986

Past!Luke tried to be sloppy and not use the table and column quoting when
inlining the AREL to SQL for attributes.  Without quoting any usage of
the references table throws a PG Error because REFERENCES is a reserved
word in PG as part of the foreign key constraint syntax.