Fix aliases operator names in specs
MSP-8986

Some of the deprecated operators besides that already exclude :os are
fully aliased, so when they were picked, they wouldn't lead to only one
operator matching.  They were passing the original check because the
unique prefixes was calculated based on the value of the operator_names
let instead of the unique prefixes across all operator names, which is
the correct way to prevent more than one operator matching.