Metasploit::Framework::Module::Class::Load::Base#metasploit_class timing
MSP-8988

Use ActiveSupport::Notifications to time
Metasploit::Framework::Module::Class:Load::Base#metasploit_class timing
when the metasploit_class constant is already in memory and when it
needs to be loaded into memory by loading the module ancestors into
memory first.  This will allow compare `use` with and without the cache
building happening in the same run of msfconsole.