Reconnect to IPC share
The SMB username parser appears to really want you to be on the IPC$
share in order to do its domain magic. This ensures that by the time you
get to reporting, your most recent connection is, in fact, on IPC$ and
not admin$.

TODO: Actually fix this upstream. In the meantime, this should resolve;
I imagine this isn't the only place where the assumption is that you're
on the IPC$ share, since it's hardcoded and not exposed as an option.