Gemify Meterpreter binaries
This removes the *.dll, *.lso and *.bin files from the data/meterpreter
folder and moves them into a gem. The code that referenced those things
directly via the data_directory configuration setting have been updated
to use the new gem, called `meterpreter_binaries` which can be found in
the meterpreter_binaries repo.

I didn't move the exe, py, jar, php, etc over, just the Windows and
POSIX binaries given that 1) they're the most turbulent, 2) they're the
biggest and 3) they're binary.

To move all the binary stuff over to this (including jar files) we'll have
to get a bit smarter with the structure (thanks to the likes of android).