Detect and clean up Thread leaks
Threads spawned by Msf::Framework#initialize were not being cleaned up,
which could cause the tests to slow down, so explicitly kill to stop the
threads and then join them to ensure they aren't left as dead in the
Thread.list.  Added around(:each) to spec_helper.rb to monitor for
thread leaks in the future: if a thread is leaked, then the example that
leaked the thread will fail.