Invalid default database name.
When calling create_db() via an RPC method, the default database name used is the name of the user running the msfrpcd process. Unfortunately, this database usually does not exist and an exception is raised resulting in a failed state. This patch forces a connection to the default postgres database; mimicking the behaviour of the postgres CLI, psql.