Minor tweak to support Linux shell upgrade to Meterpreter
The change allows the stager to be launched, backgrounded, and then the cleanup commands still run.  This was needed during the shell upgrade process.  If the stager isn't backgrounded then it continues to run and tie up the original session while the new Meterpreter session is still open.  This isn't obvious when trying to use the original 'vanilla' linux shell, it just appears unresponsive until the Meterpreter session is exited.  

The reason that this was added to the current line instead of on its own is that we need to have the ampersand '&' in that position.  You cannot place it at the end of the original command because when the command is executed it will result in the following:

/tmp/dDufuf & ; rm -f /tmp/ dDufuf

The ampersand cannot appear just before semicolon as they both serve as command terminators and it will generate an error and break the stager execution.


New version:
The added side benefit