Allow declaring root via env var in gemspecs
The bundle build tasks temporarily relocate Gemfile and gemspecs to
temporary directories when packing bundles.  For this to work, we can't
depend on the Gemfile/gemspec exisiting in a specific location relative
to other code.

This allows specifying the project root via the MSP_ROOT environment
variable.  If you do not specify MSP_ROOT, the behavior is unchanged.

MSP-10684