Revert "Fix a crash when we can't connect to PG"
This reverts commit b6deb6a342451e3628f5093bf1a3254500cc7ca3, which was
landed in PR #3673 and summarily busted all CI builds.

Obviously, this is a nonoptimal fix for this new error, but while a
proper fix is being worked on, I'd like to be able to build
successfully.

When the problem introduced by #3673 is fixed, feel free to revert this
commit and then apply the fix.