Persistence with schtasks.
This Module will create a schtask that can either run once the user is idle,
or if the user is part of the admin group you can then create schtasks that
can work at boot or user login.  Windows checks a users idle time ever 15 minutes.
This should help keep persistence on a box as a normal user.