Fix #4259 - Failure to parse ANSI codes for msf prompt on Winodws
On Windows, the msf prompt is unable to parse and render the color
codes. It also causes the prompt to be almost unusable, as in you
can't really see what you type. So this fix will not use ANSI
codes for the default prompt.

For users:
If you see stuff like <[ 4m<-[0m<-[36mmsf-pro on your console ....
know that this fix does not actually fix that. Instead, that is a
Pro-specific prompt (not framework), so you will have to wait for
Pro's fix to merge.

Fix #4259.