Fix #4304 - Blank password being tried when USER_AS_PASS is true
This is kind of a conflict between the BLANK_PASSWORDS option vs
the SMB_PASS option.

If we allow BLANK_PASSWORDS, it implies we allow the user to decide
whether there should be an empty string as the SMB password or not.
But if we always set the default value of SMB_PASS to an empty string,
we're actually not respecting the BLANK_PASSWORDS option. In other
words, as long as the BLANK_PASSWORDS option exists, it is important
to know SMB_PASS may be nil, not always a string.

Fix #4304