Known private key modules should report success
I noticed that the special-purpose "known private key" style modules do
not rely on the normal LoginScanners, so they are not reporting valid
credentials even when they have them. Also, they're aux modules, so they
do not report_vuln by default when a session is opened. This fixes at
least the second part. It'd be nice to pick up the first part, but the
references cited should be enough to go on to recover the private key if
it's desired (they're not very useful, though, outside of these specific
vulnerability contexts).