Switch the go_pro advice to something less faily
As described on #5018, on Kali 1.1.0a and thereabouts, if you type
`service metasploit start` twice in quick succession, you end up horking
up the startup process. `service metasploit restart`, turns out, tends
to be more reliable.

All this fix does is alert the user to a better mechanism to solve this.

Note, this doesn't fix completely #5018. It's merely a documentation
change. A proper fix will need to happen in release engineering and the
init scripts responsible for wrapping the service.

[See #5018]